+++ extract_msg/_rtf/tokenize_rtf.py
@@ -54,8 +54,10 @@ _KNOWN_DESTINATIONS = (
 def _finishTag(startText : bytes, reader : io.BytesIO) -> Tuple[bytes, Optional[bytes], Optional[int], bytes]:
     """
     Finishes reading a tag, returning the needed parameters to make it a
-    token. The return is a 4 tuple of the raw token bytes, the name field,
-    the parameter field (as an int), and the next character after the tag.
+    token.
+
+    The return is a 4 tuple of the raw token bytes, the name field, the
+    parameter field (as an int), and the next character after the tag.
     """
     # Very simple rules here. Anything other than a letter and we change
     # state. If the next character is a hypen, check if the character after
@@ -185,11 +187,9 @@ def _readText(startChar : bytes, reader : io.BytesIO) -> Tuple[Tuple[Token, ...]
 def tokenizeRTF(data : bytes, validateStart : bool = True) -> List[Token]:
     """
     Reads in the bytes and sets the tokens list to the contents after
-    tokenizing. If tokenizing fails, the current tokens list will not be
-    changed.
+    tokenizing.
 
-    Direct references to the previous tokens list will only point to the
-    previous and not to the current one.
+    If tokenizing fails, the current tokens list will not be changed.
 
     :param validateStart: If False, does not check the first few tags. Useful
         when tokenizing a snippet rather than a document.